fix(iota-json-rpc-tests): separate tests by api
