+++ crates/iota-json-rpc-tests/tests/transaction_builder_api.rs
@@ -1,0 +1,221 @@
+// Copyright (c) 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+#[cfg(not(msim))]
+use std::str::FromStr;
+use std::{collections::BTreeMap, path::Path};
+
+use iota_json::{call_args, type_args};
+use iota_json_rpc_api::{IndexerApiClient, TransactionBuilderClient, WriteApiClient};
+use iota_json_rpc_types::{
+    IotaObjectDataOptions, IotaObjectResponseQuery, IotaTransactionBlockEffectsAPI,
+    IotaTransactionBlockResponse, IotaTransactionBlockResponseOptions, ObjectChange,
+    TransactionBlockBytes,
+};
+use iota_macros::sim_test;
+use iota_move_build::BuildConfig;
+use iota_types::{
+    base_types::{ObjectID, SequenceNumber},
+    digests::ObjectDigest,
+    gas_coin::GAS,
+    quorum_driver_types::ExecuteTransactionRequestType,
+    IOTA_FRAMEWORK_ADDRESS,
+};
+use test_cluster::TestClusterBuilder;
+
+fn assert_same_object_changes_ignoring_version_and_digest(
+    expected: Vec<ObjectChange>,
+    actual: Vec<ObjectChange>,
+) {
+    fn collect_changes_mask_version_and_digest(
+        changes: Vec<ObjectChange>,
+    ) -> BTreeMap<ObjectID, ObjectChange> {
+        changes
+            .into_iter()
+            .map(|mut change| {
+                let object_id = change.object_id();
+                // ignore the version and digest for comparison
+                change.mask_for_test(SequenceNumber::MAX, ObjectDigest::MAX);
+                (object_id, change)
+            })
+            .collect()
+    }
+    let expected = collect_changes_mask_version_and_digest(expected);
+    let actual = collect_changes_mask_version_and_digest(actual);
+    assert!(expected.keys().all(|id| actual.contains_key(id)));
+    assert!(actual.keys().all(|id| expected.contains_key(id)));
+    for (id, exp) in &expected {
+        let act = actual.get(id).unwrap();
+        assert_eq!(act, exp);
+    }
+}
+
+#[sim_test]
+async fn test_public_transfer_object() -> Result<(), anyhow::Error> {
+    let cluster = TestClusterBuilder::new().build().await;
+    let http_client = cluster.rpc_client();
+    let address = cluster.get_address_0();
+
+    let objects = http_client
+        .get_owned_objects(
+            address,
+            Some(IotaObjectResponseQuery::new_with_options(
+                IotaObjectDataOptions::new()
+                    .with_type()
+                    .with_owner()
+                    .with_previous_transaction(),
+            )),
+            None,
+            None,
+        )
+        .await?
+        .data;
+
+    let obj = objects.clone().first().unwrap().object().unwrap().object_id;
+    let gas = objects.clone().last().unwrap().object().unwrap().object_id;
+
+    let transaction_bytes: TransactionBlockBytes = http_client
+        .transfer_object(address, obj, Some(gas), 1_000_000.into(), address)
+        .await?;
+
+    let tx = cluster
+        .wallet
+        .sign_transaction(&transaction_bytes.to_data()?);
+    let (tx_bytes, signatures) = tx.to_tx_bytes_and_signatures();
+    let tx_bytes1 = tx_bytes.clone();
+    let dryrun_response = http_client.dry_run_transaction_block(tx_bytes).await?;
+
+    let tx_response: IotaTransactionBlockResponse = http_client
+        .execute_transaction_block(
+            tx_bytes1,
+            signatures,
+            Some(
+                IotaTransactionBlockResponseOptions::new()
+                    .with_effects()
+                    .with_object_changes(),
+            ),
+            Some(ExecuteTransactionRequestType::WaitForLocalExecution),
+        )
+        .await?;
+
+    assert_same_object_changes_ignoring_version_and_digest(
+        dryrun_response.object_changes,
+        tx_response.object_changes.unwrap(),
+    );
+    Ok(())
+}
+
+#[sim_test]
+async fn test_publish() -> Result<(), anyhow::Error> {
+    let cluster = TestClusterBuilder::new().build().await;
+    let http_client = cluster.rpc_client();
+    let address = cluster.get_address_0();
+
+    let objects = http_client
+        .get_owned_objects(
+            address,
+            Some(IotaObjectResponseQuery::new_with_options(
+                IotaObjectDataOptions::new()
+                    .with_type()
+                    .with_owner()
+                    .with_previous_transaction(),
+            )),
+            None,
+            None,
+        )
+        .await?;
+    let gas = objects.data.first().unwrap().object().unwrap();
+
+    let compiled_package =
+        BuildConfig::new_for_testing().build(Path::new("../../examples/move/basics"))?;
+    let compiled_modules_bytes =
+        compiled_package.get_package_base64(/* with_unpublished_deps */ false);
+    let dependencies = compiled_package.get_dependency_storage_package_ids();
+
+    let transaction_bytes: TransactionBlockBytes = http_client
+        .publish(
+            address,
+            compiled_modules_bytes,
+            dependencies,
+            Some(gas.object_id),
+            100_000_000.into(),
+        )
+        .await?;
+
+    let tx = cluster
+        .wallet
+        .sign_transaction(&transaction_bytes.to_data()?);
+    let (tx_bytes, signatures) = tx.to_tx_bytes_and_signatures();
+
+    let tx_response = http_client
+        .execute_transaction_block(
+            tx_bytes,
+            signatures,
+            Some(IotaTransactionBlockResponseOptions::new().with_effects()),
+            Some(ExecuteTransactionRequestType::WaitForLocalExecution),
+        )
+        .await?;
+    matches!(tx_response, IotaTransactionBlockResponse {effects, ..} if effects.as_ref().unwrap().created().len() == 6);
+    Ok(())
+}
+
+#[sim_test]
+async fn test_move_call() -> Result<(), anyhow::Error> {
+    let cluster = TestClusterBuilder::new().build().await;
+    let http_client = cluster.rpc_client();
+    let address = cluster.get_address_0();
+
+    let objects = http_client
+        .get_owned_objects(
+            address,
+            Some(IotaObjectResponseQuery::new_with_options(
+                IotaObjectDataOptions::new()
+                    .with_type()
+                    .with_owner()
+                    .with_previous_transaction(),
+            )),
+            None,
+            None,
+        )
+        .await?
+        .data;
+
+    let gas = objects.first().unwrap().object().unwrap();
+    let coin = &objects[1].object()?;
+
+    // now do the call
+    let package_id = ObjectID::new(IOTA_FRAMEWORK_ADDRESS.into_bytes());
+    let module = "pay".to_string();
+    let function = "split".to_string();
+
+    let transaction_bytes: TransactionBlockBytes = http_client
+        .move_call(
+            address,
+            package_id,
+            module,
+            function,
+            type_args![GAS::type_tag()]?,
+            call_args!(coin.object_id, 10)?,
+            Some(gas.object_id),
+            10_000_000.into(),
+            None,
+        )
+        .await?;
+
+    let tx = cluster
+        .wallet
+        .sign_transaction(&transaction_bytes.to_data()?);
+
+    let (tx_bytes, signatures) = tx.to_tx_bytes_and_signatures();
+
+    let tx_response = http_client
+        .execute_transaction_block(
+            tx_bytes,
+            signatures,
+            Some(IotaTransactionBlockResponseOptions::new().with_effects()),
+            Some(ExecuteTransactionRequestType::WaitForLocalExecution),
+        )
+        .await?;
+    matches!(tx_response, IotaTransactionBlockResponse {effects, ..} if effects.as_ref().unwrap().created().len() == 1);
+    Ok(())
+}
